+++ src/utils/useTests.ts
@@ -1,14 +1,6 @@
 import { useCallback, useState } from "react";
 
-import {
-  BrowserInfo,
-  ComparisonResult,
-  TestFieldsFragment,
-  TestResult,
-  TestStatus,
-  ViewportInfo,
-} from "../gql/graphql";
-import { aggregateResult } from "./aggregateResult";
+import { BrowserInfo, TestFieldsFragment, ViewportInfo } from "../gql/graphql";
 
 type ViewportData = Pick<ViewportInfo, "id" | "name">;
 type BrowserData = Pick<BrowserInfo, "id" | "key" | "name">;
@@ -43,82 +35,10 @@ export function useTests(tests: TestFieldsFragment[]) {
     (browser: BrowserData) => onSelectBrowserId(browser.id),
     [onSelectBrowserId]
   );
-
-  const {
-    isInProgress,
-    changeCount,
-    brokenCount,
-    resultsByBrowser,
-    resultsByViewport,
-    viewportInfoById,
-  } = tests.reduce<{
-    isInProgress: boolean;
-    changeCount: number;
-    brokenCount: number;
-    resultsByBrowser: Record<string, ComparisonResult>;
-    resultsByViewport: Record<string, ComparisonResult>;
-    viewportInfoById: Record<string, TestFieldsFragment["parameters"]["viewport"]>;
-  }>(
-    (acc, test) => {
-      if (test.status === TestStatus.InProgress) {
-        acc.isInProgress = true;
-      }
-      if ([TestResult.Changed, TestResult.Added].includes(test.result)) {
-        acc.changeCount += 1;
-      }
-      if ([TestResult.CaptureError, TestResult.SystemError].includes(test.result)) {
-        acc.brokenCount += 1;
-      }
-      test.comparisons?.forEach(({ browser, result }) => {
-        acc.resultsByBrowser[browser.id] = aggregateResult([
-          result,
-          acc.resultsByBrowser[browser.id],
-        ]);
-      });
-      test.comparisons?.forEach(({ viewport, result }) => {
-        acc.resultsByViewport[viewport.id] = aggregateResult([
-          result,
-          acc.resultsByViewport[viewport.id],
-        ]);
-      });
-      acc.viewportInfoById[test.parameters.viewport.id] = test.parameters.viewport;
-      return acc;
-    },
-    {
-      isInProgress: false,
-      changeCount: 0,
-      brokenCount: 0,
-      resultsByBrowser: {},
-      resultsByViewport: {},
-      viewportInfoById: {},
-    }
-  );
-
-  // All tests have the same browsers so we can just look at the first
-  // TODO -- it's inefficient to get this data on every test when we only need the first.
-  // Instead, let's just get it on the build and pass it in
-  const browserInfoById = Object.fromEntries(
-    tests[0].comparisons.map((c) => [c.browser.id, c.browser])
-  );
-
-  const browserResults = Object.entries(resultsByBrowser).map(([id, result]) => ({
-    browser: browserInfoById[id],
-    result,
-  }));
-  const viewportResults = Object.entries(resultsByViewport).map(([id, result]) => ({
-    viewport: viewportInfoById[id],
-    result,
-  }));
-
   return {
     selectedTest,
     selectedComparison,
-    isInProgress,
-    changeCount,
-    brokenCount,
-    browserResults,
     onSelectBrowser,
-    viewportResults,
     onSelectViewport,
   };
 }